Examples/requirements.txt
python-dotenv==1.0.0
wakeonlan==3.0.0